website-project/requirements.txt
Django>=4.0,<5.0
mysqlclient>=2.0
gunicorn>=20.0
djangorestframework
django-cors-headers
pillow
djangorestframework-simplejwt